requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pandas
python-dotenv
pytest
httpx
certifi